+++ src/main/kotlin/org/nemwiz/jiracommitmessage/services/MyApplicationService.kt
@@ -1,0 +1,9 @@
+package org.nemwiz.jiracommitmessage.services
+
+
+class MyApplicationService {
+
+    init {
+        println("Application service")
+    }
+}