+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 <!-- Keep a Changelog guide -> https://keepachangelog.com -->
 
-# jira-commit-message-intellij-plugin Changelog
+# jira-commit-message Changelog
 
-## [Unreleased]
+## [2.0.0]
 ### Added
-- Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+- Support up until IntelliJ version 2020.2
+- Migrated project to [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)